+++ app/javascript/controllers/image_preview_controller.js
@@ -1,0 +1,21 @@
+import { Controller } from "@hotwired/stimulus"
+
+export default class extends Controller {
+  static targets = ["input", "preview"]
+
+  preview() {
+    const file = this.inputTarget.files[0]
+    if (!file) {
+      this.previewTarget.src = ""
+      this.previewTarget.classList.add("hidden")
+      return
+    }
+
+    const reader = new FileReader()
+    reader.onload = (e) => {
+      this.previewTarget.src = e.target.result
+      this.previewTarget.classList.remove("hidden")
+    }
+    reader.readAsDataURL(file)
+  }
+}
