+++ app/javascript/controllers/index.js
@@ -1,10 +1,6 @@
 import { application } from "./application"
-
-import HelloController from "./hello_controller"
-application.register("hello", HelloController)
-
-import PreviewController from "./preview_controller"
-application.register("preview", PreviewController)
+import ImagePreviewController from "./image_preview_controller"
+application.register("image-preview", ImagePreviewController)
 
 import TutorialModalController from "./tutorial_modal_controller"
 application.register("tutorial-modal", TutorialModalController)
