+++ app/views/boards/new.html.slim
@@ -1,5 +1,4 @@
 - content_for(:title, t('.title'))
-
 div.min-h-screen.bg-gradient-to-br.from-blue-50.to-blue-100.flex.items-center.justify-center.px-4.py-12
   .bg-white.shadow-2xl.rounded-2xl.p-8.max-w-2xl.w-full.animate-fade-in-up
     h1.text-3xl.md:text-4xl.font-bold.text-gray-800.mb-8.text-center
