+++ app/views/dogs/_form.html.slim
@@ -1,23 +1,23 @@
-div data-controller="preview"
-= form_with model: dog, local: true, html: { class: "space-y-6" } do |f|
-  div
-    = f.label :name, class: "block text-gray-700 font-semibold mb-2"
-    = f.text_field :name, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
+= form_with model: dog, local: true, html: { data: { controller: "image-preview" } } do |f|
 
-  div
-    = f.label :breed, class: "block text-gray-700 font-semibold mb-2"
-    = f.text_field :breed, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
+  = f.label :name, class: "block text-gray-700 font-semibold mb-2"
+  = f.text_field :name, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
 
-  div
-    = f.label :age, class: "block text-gray-700 font-semibold mb-2"
-    = f.number_field :age, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
 
-  div
-    = f.label :place, class: "block text-gray-700 font-semibold mb-2"
-    = f.text_field :place, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
+  = f.label :breed, class: "block text-gray-700 font-semibold mb-2"
+  = f.text_field :breed, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
 
-  div
-    = f.label :profile_image, class: "block text-gray-700 font-semibold mb-2"
-    = f.file_field :profile_image, class: "mt-1 block w-full text-gray-600"
+
+  = f.label :age, class: "block text-gray-700 font-semibold mb-2"
+  = f.number_field :age, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
+
+
+  = f.label :place, class: "block text-gray-700 font-semibold mb-2"
+  = f.text_field :place, class: "mt-1 block w-full border border-gray-300 rounded-lg shadow-sm px-3 py-2 focus:ring-blue-400 focus:border-blue-400"
+
+
+  = f.label :profile_image, class: "block text-gray-700 font-semibold mb-2"
+  = f.file_field :profile_image, data: { action: "image-preview#preview", image_preview_target: "input" }
+  img.hidden.mt-4.max-w-xs.rounded-lg.shadow-lg data-image-preview-target="preview"
 
   = f.submit nil, class: "w-full bg-blue-600 hover:bg-blue-700 text-white font-semibold py-2 rounded-lg transition duration-300"
