+++ app/views/dogs/show.html.slim
@@ -26,4 +26,4 @@ div.min-h-screen.bg-gradient-to-br.from-yellow-50.to-pink-100.flex.items-center.
     .mt-8.flex.justify-center.gap-4
       = link_to "編集", edit_dog_path(@dog), class: "bg-blue-500 hover:bg-blue-600 text-white font-semibold px-5 py-2 rounded-lg shadow transition duration-300"
       = link_to "一覧に戻る", dogs_path, class: "bg-gray-300 hover:bg-gray-400 text-gray-800 font-semibold px-5 py-2 rounded-lg shadow transition duration-300"
-      = link_to "削除", dogs_path, class: "bg-gray-300 hover:bg-gray-400 text-gray-800 font-semibold px-5 py-2 rounded-lg shadow transition duration-300"
+      = link_to "削除", dogs_path, class: "inline-flex items-center bg-red-500 hover:bg-red-600 text-white font-semibold py-2 px-4 rounded-lg transition-all duration-300"
